Renaming collection and added some improvements
