+++ Split/Common/Structs/ConcurrentArrayQueue.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 /// A thread-safe array.
-class SynchronizedArrayQueue<T> {
+class ConcurrentArrayQueue<T> {
     private let queue = DispatchQueue(label: "io.Split.Structs.ArraySerialBlockingQueue", attributes: .concurrent)
     private var array = [T]()
     private var firstAppend = true
@@ -27,7 +27,10 @@ class SynchronizedArrayQueue<T> {
         var element: T?
         queue.sync {
             if !array.isEmpty {
-                element = self.array.remove(at: 0)
+                queue.async(flags: .barrier) { [weak self] in
+                    guard let self = self else { return }
+                    element = self.array.remove(at: 0)
+                }
             }
         }
         return element
