+++ Split/Common/Structs/SynchronizedDictionarySet.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-class ConcurrentDictionarySet<K: Hashable, T: Hashable> {
+class SynchronizedDictionarySet<K: Hashable, T: Hashable> {
 
     private var queue: DispatchQueue = DispatchQueue(label: "split-concurrent-dictionary-map",
                                                      target: .global())
